Implement UDPM message sending
Also consolidated the UDPM into a single file. The backend and the
frontend are in the same module and the datagram module has been
removed.
